removes parse runner class replacing all calls with parse interface.
